backend/requirements.txt
fastapi
uvicorn
pydantic-ai
openai
python-dotenv
